+++ djangocms_alias/cms_config.py
@@ -13,6 +13,14 @@ try:
 except (ImportError, LookupError):
     AliasContentConfig = None
 
+try:
+    from djangocms_versioning.constants import DRAFT  # NOQA
+
+    djangocms_versioning_installed = True
+except ImportError:
+    djangocms_versioning_installed = False
+
+
 class AliasCMSConfig(CMSAppConfig):
     cms_enabled = True
     cms_toolbar_enabled_models = [(AliasContent, render_alias_content)]
@@ -22,7 +30,9 @@ class AliasCMSConfig(CMSAppConfig):
         settings, 'MODERATING_ALIAS_MODELS_ENABLED', True)
     djangocms_versioning_enabled = getattr(
         settings, 'VERSIONING_ALIAS_MODELS_ENABLED', True)
-    if djangocms_versioning_enabled:
+
+    if djangocms_versioning_enabled and djangocms_versioning_installed:
+
         from cms.utils.i18n import get_language_tuple
 
         from djangocms_versioning.datastructures import VersionableItem