+++ CHANGELOG.rst
@@ -4,6 +4,7 @@ Changelog
 
 Unreleased
 ==========
+* feat: Add versioning actions to Alias AliasContent Manager
 * feat: Github Actions integration
 * Python 3.8, 3.9 support added
 * Django 3.0, 3.1 and 3.2 support added